// src/app/model/mandate.model.ts
export class Mandate  {  

    acct_no: string;
    acct_type: string;
    last_name: string;
    otherNames: string;
    address_1: string;
    address_2: string;
    address_3: string;
    branch_no: string;
    branch_name: string;
    no_Signatures: string;
    signatory: string;
    image_description: string;
    image_id: string;
    mandateType: string;
    Photo: string;
    create_dt: string;
    photoBase64: string;
}